add new class, changed and rename methods. Ref #263
