+++ app/go/go_directory_management.py
@@ -1,0 +1,20 @@
+import os, shutil
+
+class GoDirectoryManagement:
+	def __init__(self, path=None):
+		self.path = path
+
+	def get_path(self):
+		return self.path
+
+	def set_path(self, path):
+		self.path = path
+
+	def create_dir(self):
+		os.makedirs(self.get_path())
+
+	def is_dir(self):
+		return os.path.isdir(self.get_path())
+
+	def remove_dir(self):
+		shutil.rmtree(self.get_path())
